AWS stub helper to support dir
