+++ spec/support/aws_stub_helper.rb
@@ -1,0 +1,16 @@
+module AwsStubHelper
+  def build_stub_aws_client(klass, api_request_log = [])
+    client = klass.new(
+      region: Broadside.config.aws.region,
+      credentials: Aws::Credentials.new('access', 'secret'),
+      stub_responses: true
+    )
+
+    client.handle do |context|
+      api_request_log << { context.operation_name => context.params }
+      @handler.call(context)
+    end
+
+    client
+  end
+end
